refactor(main): 重构代码，分离IPC函数
新增文件ipchandler，封装所有负责ipc的函数，并添加详细注释
