+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "TimeFlow",
   "private": true,
-  "version": "0.0.1",
+  "version": "0.0.89",
   "type": "module",
   "main": "./main.js",
   "scripts": {
